+++ driver/result.go
@@ -18,15 +18,13 @@ var (
 	_ driver.RowsColumnTypeNullable         = (*queryResult)(nil)
 	_ driver.RowsColumnTypePrecisionScale   = (*queryResult)(nil)
 	_ driver.RowsColumnTypeScanType         = (*queryResult)(nil)
-	_ driver.RowsNextResultSet              = (*queryResult)(nil)
+	/*
+		currently not used
+		could be implemented as pointer to next queryResult (advancing by copying data from next)
+		_ driver.RowsNextResultSet = (*queryResult)(nil)
+	*/
 
-	_ driver.Rows                           = (*callResult)(nil)
-	_ driver.RowsColumnTypeDatabaseTypeName = (*callResult)(nil)
-	_ driver.RowsColumnTypeLength           = (*callResult)(nil)
-	_ driver.RowsColumnTypeNullable         = (*callResult)(nil)
-	_ driver.RowsColumnTypePrecisionScale   = (*callResult)(nil)
-	_ driver.RowsColumnTypeScanType         = (*callResult)(nil)
-	_ driver.RowsNextResultSet              = (*callResult)(nil)
+	_ driver.Rows = (*callResult)(nil)
 )
 
 type prepareResult struct {
@@ -41,12 +39,6 @@ func (pr *prepareResult) isProcedureCall() bool { return pr.fc.IsProcedureCall()
 
 // numField returns the number of parameter fields in a database statement.
 func (pr *prepareResult) numField() int { return len(pr.parameterFields) }
-
-// onCloser defines getter and setter for a function which should be called when closing.
-type onCloser interface {
-	onClose() func()
-	setOnClose(func())
-}
 
 // NoResult is the driver.Rows drop-in replacement if driver Query or QueryRow is used for statements that do not return rows.
 var noResult = new(noResultType)
@@ -70,16 +62,8 @@ type queryResult struct {
 	conn         *conn
 	rsID         uint64
 	pos          int
-	_onClose     func()
 	attributes   p.PartAttributes
-	closed       bool
 }
-
-// onClose implements the onCloser interface
-func (qr *queryResult) onClose() func() { return qr._onClose }
-
-// setOnClose implements the onCloser interface
-func (qr *queryResult) setOnClose(f func()) { qr._onClose = f }
 
 // Columns implements the driver.Rows interface.
 func (qr *queryResult) Columns() []string {
@@ -95,11 +79,6 @@ func (qr *queryResult) Columns() []string {
 
 // Close implements the driver.Rows interface.
 func (qr *queryResult) Close() error {
-	if !qr.closed && qr._onClose != nil {
-		defer qr._onClose()
-	}
-	qr.closed = true
-
 	if qr.attributes.ResultsetClosed() {
 		return nil
 	}
@@ -171,35 +150,14 @@ func (qr *queryResult) ColumnTypeScanType(idx int) reflect.Type {
 	return qr.fields[idx].ScanType()
 }
 
-/*
-driver.RowsNextResultSet:
-- currently not used
-- could be implemented as pointer to next queryResult (advancing by copying data from next)
-*/
-
-// HasNextResultSet implements the driver.RowsNextResultSet interface.
-func (qr *queryResult) HasNextResultSet() bool { return false }
-
-// NextResultSet implements the driver.RowsNextResultSet interface.
-func (qr *queryResult) NextResultSet() error { return io.EOF }
-
 type callResult struct { // call output parameters
 	conn         *conn
 	outputFields []*p.ParameterField
 	fieldValues  []driver.Value
 	decodeErrors p.DecodeErrors
 	_columns     []string
-	qrs          []*queryResult // table output parameters
 	eof          bool
-	closed       bool
-	_onClose     func()
 }
-
-// onClose implements the onCloser interface
-func (cr *callResult) onClose() func() { return cr._onClose }
-
-// setOnClose implements the onCloser interface
-func (cr *callResult) setOnClose(f func()) { cr._onClose = f }
 
 // Columns implements the driver.Rows interface.
 func (cr *callResult) Columns() []string {
@@ -213,7 +171,7 @@ func (cr *callResult) Columns() []string {
 	return cr._columns
 }
 
-// / Next implements the driver.Rows interface.
+// Next implements the driver.Rows interface.
 func (cr *callResult) Next(dest []driver.Value) error {
 	if len(cr.fieldValues) == 0 || cr.eof {
 		return io.EOF
@@ -231,48 +189,4 @@ func (cr *callResult) Next(dest []driver.Value) error {
 }
 
 // Close implements the driver.Rows interface.
-func (cr *callResult) Close() error {
-	if !cr.closed && cr._onClose != nil {
-		cr._onClose()
-	}
-	cr.closed = true
-	return nil
-}
-
-// ColumnTypeDatabaseTypeName implements the driver.RowsColumnTypeDatabaseTypeName interface.
-func (cr *callResult) ColumnTypeDatabaseTypeName(idx int) string {
-	return cr.outputFields[idx].TypeName()
-}
-
-// ColumnTypeLength implements the driver.RowsColumnTypeLength interface.
-func (cr *callResult) ColumnTypeLength(idx int) (int64, bool) {
-	return cr.outputFields[idx].TypeLength()
-}
-
-// ColumnTypeNullable implements the driver.RowsColumnTypeNullable interface.
-func (cr *callResult) ColumnTypeNullable(idx int) (bool, bool) {
-	return cr.outputFields[idx].Nullable(), true
-}
-
-// ColumnTypePrecisionScale implements the driver.RowsColumnTypePrecisionScale interface.
-func (cr *callResult) ColumnTypePrecisionScale(idx int) (int64, int64, bool) {
-	return cr.outputFields[idx].TypePrecisionScale()
-}
-
-// ColumnTypeScanType implements the driver.RowsColumnTypeScanType interface.
-func (cr *callResult) ColumnTypeScanType(idx int) reflect.Type {
-	return cr.outputFields[idx].ScanType()
-}
-
-// HasNextResultSet implements the driver.RowsNextResultSet interface.
-func (cr *callResult) HasNextResultSet() bool { return false }
-
-// NextResultSet implements the driver.RowsNextResultSet interface.
-func (cr *callResult) NextResultSet() error { return io.EOF }
-
-func (cr *callResult) appendTableRowsFields() {
-	for i, qr := range cr.qrs {
-		cr.outputFields = append(cr.outputFields, p.NewTableRowsParameterField(i))
-		cr.fieldValues = append(cr.fieldValues, qr)
-	}
-}
+func (cr *callResult) Close() error { return nil }